general_react_agent/tavily_search_client.py
from tavily import TavilyClient
from dotenv import load_dotenv
import os

load_dotenv()

tavily_client = TavilyClient(api_key = os.getenv("TAVILY_API_KEY"))

response = tavily_client.get_search_context("who is leo messi?") 
print(response)