pyproject.toml
[tool.poetry]
name = "general-react-agent"
version = "0.1.0"
description = ""
authors = ["Nimesh Monson <nimesh.monson.96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
openai = "^1.58.1"
tavily-python = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
